+++ t-rex-core/src/core/config.rs
@@ -7,9 +7,11 @@ use crate::core::gridcfg::ExtentCfg;
 use regex::Regex;
 use serde::Deserialize;
 use std;
+use std::collections::HashMap;
 use std::env;
 use std::fs::File;
 use std::io::prelude::*;
+use tera::{Context, Tera};
 use toml::Value;
 
 pub trait Config<'a, C: Deserialize<'a>>
@@ -245,17 +247,27 @@ pub fn read_config<'a, T: Deserialize<'a>>(path: &str) -> Result<T, String> {
 
 /// Parse the configuration into an config struct.
 pub fn parse_config<'a, T: Deserialize<'a>>(config_toml: String, path: &str) -> Result<T, String> {
-    // Replace variables ${var}
+    // Check for old ${var} expressions
     let re = Regex::new(r"\$\{([[:alnum:]]+)\}").unwrap();
-    let toml = re.replace_all(&config_toml, |cap: &regex::Captures| {
-        match env::var(&cap[1]) {
-            Ok(val) => val,
-            Err(_) => cap[0].to_string(),
-        }
-    });
-    if let Some(cap) = re.captures(&toml) {
-        return Err(format!("Environment variable `{}` undefined", &cap[1]));
+    if re.is_match(&config_toml) {
+        return Err(
+            "Replace old environment variable syntax ${VARNAME} with `{{env.VARNAME}}`".to_string(),
+        );
     }
+
+    // Parse template
+    let mut tera = Tera::default();
+    tera.add_raw_template(path, &config_toml)
+        .map_err(|e| format!("Template error: {}", e))?;
+    let mut context = Context::new();
+    let mut env = HashMap::new();
+    for (key, value) in env::vars() {
+        env.insert(key, value);
+    }
+    context.insert("env", &env);
+    let toml = tera
+        .render(path, &context)
+        .map_err(|e| format!("Template error: {}", e))?;
 
     toml.parse::<Value>()
         .and_then(|cfg| cfg.try_into::<T>())